moved: tetra/client.py and tetra/protos/*
