+++ tetra/__init__.py
@@ -4,7 +4,7 @@ setup_logging()
 from . import remote_execution_pb2, remote_execution_pb2_grpc
 from .core.resources.resource_manager import ResourceManager
 from .core.resources.serverless import ServerlessResource
-from .core.client.client_manager import remote
+from .client import remote
 
 __all__ = [
     "remote",
